requirements.txt
requests>=2.8.1
python-magic==0.4.27
python-dotenv==0.20.0